Move MockContextLifecycle to scheduler mod
Moves the MockContextLifecycle from the serial scheduler's tests module
to the top-level scheduler's tests module. This allows the mock struct
to be used by other scheduler implementations' tests.
